+++ backend/src/middleware/auth.middleware.ts
@@ -1,4 +1,4 @@
-import { getClerkClient } from '../clerk';
+import { getClerkClient } from '../auth/clerk';
 
 import { createMiddleware } from 'hono/factory';
 import { Bindings, Variables } from 'hono/types';
